requirements.txt
anyio==4.11.0
async-timeout==5.0.1
betterproto==2.0.0b7
certifi==2025.11.12
charset-normalizer==3.4.4
ffmpeg-python==0.2.0
ffmpy==1.0.0
filelock==3.20.0
fsspec==2025.10.0
future==1.0.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
mashumaro==3.17
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
numba==0.62.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
protobuf==6.33.1
protobuf3-to-dict==0.1.5
pyee==13.0.0
python-dateutil==2.9.0.post0
python-socks==2.7.3
regex==2025.11.3
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.12.0
TikTokLive==6.6.5
torch==2.9.1
tqdm==4.67.1
triton==3.5.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
websockets_proxy==0.1.3
